+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds `embedCode` prop to `VideoPlayer` component.
+
 ## 0.25.6 (December 16, 2021)
 
 ### Adds